chore: added get_health and get_account functions
