+++ src/types.ts
@@ -1,5 +1,6 @@
 import {
     ABI,
+    API,
     BlockTimestamp,
     Checksum256,
     Float64,
@@ -201,5 +202,90 @@ export namespace V2 {
         @Struct.field('bool') declare cached_lib: boolean
         @Struct.field(Action, {array: true}) declare actions: Action[]
     }
-    
+    @Struct.type('get_account_response_link')
+    export class GetAccountResponseLink extends Struct {
+        @Struct.field(BlockTimestamp) declare timestamp: BlockTimestamp
+        @Struct.field(Name) declare permission: Name
+        @Struct.field(Name) declare code: Name
+        @Struct.field(Name) declare action: Name
+    }
+
+    @Struct.type('get_account_response')
+    export class GetAccountResponse extends Struct {
+        @Struct.field(Float64) declare query_time_ms: Float64
+        @Struct.field(API.v1.AccountObject) declare account: API.v1.AccountObject
+        @Struct.field(GetAccountResponseLink, {array: true}) declare links: GetAccountResponseLink[]
+        @Struct.field(TokenInfo, {array: true}) declare tokens: TokenInfo[]
+        @Struct.field(UInt64) declare total_actions: UInt64
+        @Struct.field(Action, { array: true }) declare actions: Action[]
+    }
+
+    // For the `health` field
+    @Struct.type('health_entry')
+    export class HealthEntry extends Struct {
+        @Struct.field('string') declare service: string;
+        @Struct.field('string') declare status: string;
+        @Struct.field('map') declare service_data?: NodeosRPCServiceData | ElasticsearchServiceData;
+        @Struct.field(UInt64) declare time: UInt64;
+    }
+
+    @Struct.type('nodeos_rpc_service_data')
+    export class NodeosRPCServiceData extends Struct {
+        @Struct.field(UInt64) declare head_block_num: UInt64;
+        @Struct.field('string') declare head_block_time: string;
+        @Struct.field(UInt64) declare time_offset: UInt64;
+        @Struct.field(UInt64) declare last_irreversible_block: UInt64;
+        @Struct.field('string') declare chain_id: string;
+    }
+
+    @Struct.type('elasticsearch_service_data')
+    export class ElasticsearchServiceData extends Struct {
+        @Struct.field('string') declare active_shards: string;
+        @Struct.field(UInt64) declare head_offset: UInt64;
+        @Struct.field(UInt64) declare first_indexed_block: UInt64;
+        @Struct.field(UInt64) declare last_indexed_block: UInt64;
+        @Struct.field(UInt64) declare total_indexed_blocks: UInt64;
+        @Struct.field(UInt64) declare missing_blocks: UInt64;
+        @Struct.field('string') declare missing_pct: string;
+    }
+
+    // For the `features` field
+    @Struct.type('streaming_features')
+    export class StreamingFeatures extends Struct {
+        @Struct.field('bool') declare enable: boolean;
+        @Struct.field('bool') declare traces: boolean;
+        @Struct.field('bool') declare deltas: boolean;
+    }
+
+    @Struct.type('tables_features')
+    export class TablesFeatures extends Struct {
+        @Struct.field('bool') declare proposals: boolean;
+        @Struct.field('bool') declare accounts: boolean;
+        @Struct.field('bool') declare voters: boolean;
+    }
+
+    // For the `features` in the main type
+    @Struct.type('features')
+    export class Features extends Struct {
+        @Struct.field(StreamingFeatures) declare streaming: StreamingFeatures;
+        @Struct.field(TablesFeatures) declare tables: TablesFeatures;
+        @Struct.field('bool') declare index_deltas: boolean;
+        @Struct.field('bool') declare index_transfer_memo: boolean;
+        @Struct.field('bool') declare index_all_deltas: boolean;
+        @Struct.field('bool') declare deferred_trx: boolean;
+        @Struct.field('bool') declare failed_trx: boolean;
+        @Struct.field('bool') declare resource_limits: boolean;
+        @Struct.field('bool') declare resource_usage: boolean;
+    }
+
+    // The main type
+    @Struct.type('get_health_response')
+    export class GetHealthResponse extends Struct {
+        @Struct.field('string') declare version: string;
+        @Struct.field('string') declare version_hash: string;
+        @Struct.field('string') declare host: string;
+        @Struct.field(HealthEntry, { array: true }) declare health: HealthEntry[];
+        @Struct.field(Features) declare features: Features;
+        @Struct.field(Float64) declare query_time_ms: Float64;
+    }
 }
